refactor: separate login handling logic into separate files
